+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -23,7 +23,7 @@ import { PaysDeLaLoire2024 } from './20240101_PaysDeLaLoire';
 import { PetrLunevilloisS12023 } from './20240108_PetrLunevillois';
 import { Pmgf } from './20221102_PMGF';
 import { Pmgf2023 } from './20230502_PMGF';
-import { PmgfLate2023 } from './PmgfLate2023';
+import { PmgfLate2023 } from './20231001_PMGF';
 import { SMT2022 } from './SMT2022';
 import { SMT2023 } from './SMT2023';
 import { SMTC2024 } from './SMTC2024';
